requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
shap==0.46.0
matplotlib==3.9.1
joblib==1.4.2
python-dotenv==1.0.1
reportlab
bcrypt>=4.1.0